Changed cookie maxAge from int to long as suggested by @Mjl33 (#1494)

+++ modules/spring-mock-mvc/src/main/java/io/restassured/module/mockmvc/internal/MockMvcRequestSenderImpl.java
@@ -370,7 +370,7 @@ class MockMvcRequestSenderImpl implements MockMvcRequestSender, MockMvcRequestAs
                     servletCookie.setDomain(cookie.getDomain());
                 }
                 if (cookie.hasMaxAge()) {
-                    servletCookie.setMaxAge(cookie.getMaxAge());
+                    servletCookie.setMaxAge((int) cookie.getMaxAge());
                 }
                 if (cookie.hasPath()) {
                     servletCookie.setPath(cookie.getPath());
